Fix TrackId to preserve documented -1 unknown value (#536)
## Summary

The spec defines `m_trackId` as `int8` with `-1` meaning unknown, but
`ISessionDataBase.TrackId` was declared as `ushort` and
`PacketToSessionData` read it as an unsigned byte, so the documented
`-1` silently became `255`. `TrackId` is now `short` across the
interface and the `SessionData2019`–`SessionData2023` implementations,
and the parser casts through `sbyte`, matching the existing pattern used
for `ICarStatusDataBase.DRSAllowed`.

## Linked issues

Closes #59

## Review notes

`MatchTrackIdToName` now takes a `short` parameter; unmapped/negative
ids still fall through to `"Unknown"`. Added a regression test
(`GetSessionDataTrackIdByteValue255ReturnsMinusOne`) alongside the
existing int8-decoding tests in `PacketSignedByteFieldsTests.cs`.

+++ F1Server.Core/Packets/PacketToObject/PacketToSessionData.cs
@@ -231,7 +231,7 @@ internal class PacketToSessionData : PacketToXBase
     /// </summary>
     /// <param name="trackId">Id of track</param>
     /// <returns>Name of track</returns>
-    private string MatchTrackIdToName(ushort trackId)
+    private string MatchTrackIdToName(short trackId)
     {
         string trackName = trackId switch
                            {
@@ -322,7 +322,7 @@ internal class PacketToSessionData : PacketToXBase
 
             actOffset += ConstData.TypeUInt8;
 
-            sessionData.TrackId = Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
+            sessionData.TrackId = (sbyte)Unsafe.ReadUnaligned<byte>(ref Unsafe.Add(ref dataPacket, actOffset));
 
             sessionData.TrackName = MatchTrackIdToName(sessionData.TrackId);
